requirements.txt
torch
tiktoken
pyyaml
tqdm
transformers